chore: split up integration tests into separate files
