+++ cli/tests/integration/repl_tests.rs
